# kvcoord: optimize batch truncation loop