… pkg/kv/kvclient/kvcoord/batch_test.go
… pkg/kv/kvclient/kvcoord/dist_sender_server_test.go
… pkg/kv/kvclient/kvcoord/truncate_test.go