@yuzefovich 2022-06-16
Some macro benchmark numbers: here is the comparison of all TPCH queries, ran 20 times, before this PR and after this PR (with the current default settings, meaning the streamer is enabled):
```
Q1: 3.24s	3.25s	0.30%
Q2: 0.19s	0.19s	0.47%
Q3: 1.39s	1.35s	-2.86%
Q4: 1.04s	1.04s	0.29%
Q5: 1.81s	1.70s	-5.95%
Q6: 6.62s	5.18s	-21.78%
Q7: 5.50s	4.72s	-14.19%
Q8: 0.81s	0.71s	-12.11%
Q9: 4.26s	4.04s	-5.17%
Q10: 1.39s	1.34s	-3.37%
Q11: 0.51s	0.48s	-6.08%
Q12: 6.77s	5.23s	-22.71%
Q13: 1.03s	0.98s	-5.57%
Q14: 0.55s	0.44s	-19.58%
Q15: 4.24s	3.42s	-19.51%
Q16: 0.76s	0.74s	-2.20%
Q17: 0.70s	0.68s	-2.98%
Q18: 5.45s	5.61s	3.09%
Q19: 3.36s	3.33s	-0.91%
Q20: 11.18s	9.52s	-14.79%
Q21: 5.75s	4.84s	-15.81%
Q22: 0.53s	0.54s	1.70%
```